requirements.txt
python-telegram-bot==20.3
Flask==2.2.3
requests==2.28.2
matplotlib==3.7.1
sqlite3==2.6.0
python-dotenv==1.0.0
APScheduler==3.9.1
Pillow==9.5.0

pip install python-telegram-bot flask requests matplotlib python-dotenv apscheduler pillow
#requirements.txt